+++ demo/example-12/.babelrc
