+++ demo/example-12/README.md
@@ -1,0 +1,25 @@
+## 打包实践（esint）
+
+### 主要内容
+
+讲解了如何使用在webpack中使用eslint，以及使用eslint的常用方法，同时层层递进的讲解了如何配置一个自己的 `.eslintrc.js` 文件
+
+具体说明请参考：
+
+[webpack 如何引入并使用eslint](https://github.com/lvzhenbang/webpack-learning/tree/master/doc/babel-eslint.md)
+
+### 运行项目
+
+```
+npm install 安装依赖
+
+npm run dev 打包编译
+
+npm run lint 项目文件的规范检测
+```
+
+### LICENSE
+
+[MIT](https://opensource.org/licenses/MIT)
+
+Copyright (c) 2018-1-11 present, Zhenbang Lv
