+++ demo/example-12/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>babel-eslint</title>
+</head>
+<body>
+	<script src="https://code.jquery.com/jquery-1.11.3.js"></script>
+	<script src="./app.js"></script>
+</body>
+</html>
