+++ demo/example-12/main.js
@@ -1,0 +1,2 @@
+const main = require("./src/module.js");
+main();
