+++ demo/example-12/package.json
@@ -1,25 +1,37 @@
 {
-  "name": "example-11",
+  "name": "example-12",
   "version": "1.0.0",
-  "description": "this is a example-11",
+  "description": "this is a example-12",
   "main": "main.js",
   "scripts": {
-    "dev": "webpack"
+    "dev": "webpack",
+    "lint": "eslint --ext .js,.html,.vue src"
   },
   "keywords": [
     "webpack",
-    "demo"
+    "demo",
+    "eslint"
   ],
   "author": "lvzhenbang",
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "css-loader": "^0.28.8",
+    "eslint": "^4.16.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-google": "^0.9.1",
+    "eslint-config-standard": "^11.0.0-beta.0",
+    "eslint-friendly-formatter": "^3.0.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-html": "^4.0.1",
+    "eslint-plugin-standard": "^3.0.1",
     "style-loader": "^0.19.1",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.1",
