+++ demo/example-12/src/module.styl
@@ -1,0 +1,13 @@
+.my-box
+  width 200px
+  height 30px
+  border 1px solid #ccc
+  
+.icon 
+  margin 10px 0
+  font-size 42px
+  line-height 100px
+  color #333
+  transition font-size 0.25s ease-out 0s
+  &:hover 
+    font-size 100px
