+++ demo/example-12/src/template.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>template</title>
+</head>
+<body>
+
+</body>
+</html>
