+++ demo/example-12/src/template.vue
@@ -1,0 +1,10 @@
+<template>
+	<p>hello eslint</p>
+</template>
+<script>
+	var foo = 1;
+</script>
+
+<style>
+	
+</style>
