+++ demo/example-12/webpack.config.js
@@ -14,6 +14,15 @@ module.exports = {
                 test: /\.js$/,
                 use: 'babel-loader',
                 include: path.resolve(__dirname, 'src')
+            },
+            {
+                test: /\.js$/,
+                loader: 'eslint-loader',
+                enforce: "pre",
+                include: [path.resolve(__dirname, 'src')],
+                options: {
+                    formatter: require('eslint-friendly-formatter')
+                }
             }
         ]
     },
